backend/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# pandas==2.1.3
# numpy==1.26.2
# pyarrow==14.0.1
# sentence-transformers==2.2.2
# faiss-cpu==1.7.4
# python-multipart==0.0.6
# httpx==0.25.1
# sqlite3
fastapi
uvicorn
pandas
numpy
pyarrow
sentence-transformers
faiss-cpu
python-multipart
httpx
# sqlite3